requirements.txt
absl-py==1.0.0
aioice==0.7.5
aiortc==1.2.1
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.4.0
av==8.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.3
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.25
google-crc32c==1.3.0
idna==3.3
importlib-resources==5.4.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
notebook==6.4.6
numpy==1.22.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pycparser==2.21
pydeck==0.7.1
pyee==8.2.2
Pygments==2.11.2
pylibsrtp==0.7.0
Pympler==1.0.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.3.1
streamlit-webrtc==0.33.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
